PythonCore/psql/final_project/requirements.txt
#библиотека для подключения к psql

psycopg2-binary==2.9.10